* lib/ftell.c (ftell): Comment out cast.

+++ lib/ftell.c
@@ -28,7 +28,7 @@ ftell (FILE *fp)
   /* Use the replacement ftello function with all its workarounds.  */
   off_t offset = ftello (fp);
   if (LONG_MIN <= offset && offset <= LONG_MAX)
-    return (long)offset;
+    return /* (long) */ offset;
   else
     {
       errno = EOVERFLOW;
